Merge in MCUCORE/mcu-sdk-github from feature/MCUXGH-1486-github-mcuxpresso-sdk-2.15.000-rt1180-b0-rfp-merge-release-tag-back to main

+++ components/phy/device/phyrtl8201/fsl_phyrtl8201.c
@@ -21,7 +21,7 @@
 #define PHY_PAGE_INTR_ADDR  (7U)
 #define PHY_INER_REG        (19U)
 
-#define PHY_INER_LINKSTATUS_CHANGE_MASK (1U << 13)
+#define PHY_INER_LINKSTATUS_CHANGE_MASK ((uint16_t)1U << 13)
 
 /*! @brief Defines the timeout macro. */
 #define PHY_READID_TIMEOUT_COUNT 1000U
